Incorperated alex's code
