+++ src/app/RootLayoutClient.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { UserProvider } from "@/app/components/hooks/UserContext";
+import { UserProvider } from "@/components/hooks/UserContext";
 import React from "react";
 
 export default function RootLayoutClient({
