+++ src/app/admin/subject/[slug]/[unit]/[articleNumber]/page.tsx
@@ -1,12 +1,14 @@
 "use client";
 import ArticleCreator from "@/app/article-creator/_components/ArticleCreator";
-import { useUser } from "@/app/components/hooks/UserContext";
-import Footer from "@/app/components/ui/footer";
-import Navbar from "@/app/components/ui/navbar";
+import { useUser } from "@/components/hooks/UserContext";
+import Footer from "@/components/ui/footer";
+import Navbar from "@/components/ui/navbar";
 import { usePathname, useRouter } from "next/navigation";
 
 const Page = ({ params }: { params: { slug: string } }) => {
-  const { user, loading, error } = useUser();
+  const { user, loading, error, setError } = useUser();
+  // Clear errors incase they exist from other pages
+  setError(null);
   const router = useRouter();
   const pathname = usePathname();
 
@@ -17,7 +19,6 @@ const Page = ({ params }: { params: { slug: string } }) => {
       /-/g,
       " ",
     );
-
 
   if (!user || user?.access === "user") {
     router.push("/");
