+++ src/app/article-creator/_components/custom_questions/QuestionInstance.tsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from "react";
 import { QuestionFormat } from "@/types/questions";
 
-import CheckForUnderstanding from "@/app/components/questions/checkForUnderstanding";
-import QuizRenderer from "@/app/components/questions/quizRenderer";
+import CheckForUnderstanding from "@/components/questions/checkForUnderstanding";
+import QuizRenderer from "@/components/questions/quizRenderer";
 import QuestionsInputInterface from "./QuestionsInputInterface";
 
 const useSyncedQuestions = (instanceId: string) => {
